+++ packages/edit-site/src/components/sidebar-dataviews/default-views.js
@@ -39,9 +39,10 @@ const DEFAULT_POST_BASE = {
 	page: 1,
 	perPage: 20,
 	sort: {
-		field: 'date',
-		direction: 'desc',
+		field: 'title',
+		direction: 'asc',
 	},
+	showLevels: true,
 	titleField: 'title',
 	mediaField: 'featured_media',
 	fields: [ 'author', 'status' ],